Finish the project

+++ src/main/java/com/slamach/todoapi/service/TaskService.java
@@ -1,8 +1,10 @@
 package com.slamach.todoapi.service;
 
 import com.slamach.todoapi.dto.TaskDto;
+import com.slamach.todoapi.exception.PermissionDeniedException;
 import com.slamach.todoapi.exception.TaskNotFoundException;
 import com.slamach.todoapi.model.Task;
+import com.slamach.todoapi.model.User;
 import com.slamach.todoapi.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,22 +16,34 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TaskService {
   private final TaskRepository taskRepository;
+  private final UserService userService;
 
-  public List<Task> getTasks() {
-    return taskRepository.findAll();
+  public List<Task> getUserTasks(String username) {
+    User user = (User) userService.loadUserByUsername(username);
+    return taskRepository.findByOwner(user).get();
   }
 
-  public Task addTask(TaskDto taskDto) {
-    return taskRepository.save(new Task(taskDto));
+  public Task addTask(TaskDto taskDto, String username) {
+    User user = (User) userService.loadUserByUsername(username);
+    return taskRepository.save(new Task(
+        taskDto.getContent(),
+        taskDto.getCompleted(),
+        user
+    ));
   }
 
-  public Task updateTask(Long taskId, TaskDto taskDto) {
+  public Task updateTask(Long taskId, TaskDto taskDto, String username) {
     Optional<Task> foundTask = taskRepository.findById(taskId);
     if (foundTask.isEmpty()) {
       throw new TaskNotFoundException("No such task found.");
     }
 
     Task updatedTask = foundTask.get();
+    User user = (User) userService.loadUserByUsername(username);
+    if (!user.getUsername().equals(updatedTask.getOwner().getUsername())) {
+      throw new PermissionDeniedException("You are not allowed to modify this task.");
+    }
+
     if (taskDto.getContent() != null) {
       updatedTask.setContent(taskDto.getContent());
     }
@@ -40,13 +54,19 @@ public class TaskService {
     return taskRepository.save(updatedTask);
   }
 
-  public Task deleteTask(Long taskId) {
+  public Task deleteTask(Long taskId, String username) {
     Optional<Task> foundTask = taskRepository.findById(taskId);
     if (foundTask.isEmpty()) {
       throw new TaskNotFoundException("No such task found.");
     }
 
+    Task deletingTask = foundTask.get();
+    User user = (User) userService.loadUserByUsername(username);
+    if (!user.getUsername().equals(deletingTask.getOwner().getUsername())) {
+      throw new PermissionDeniedException("You are not allowed to delete this task.");
+    }
+
     taskRepository.deleteById(taskId);
-    return foundTask.get();
+    return deletingTask;
   }
 }
